+++ frontend/packages/shared/src/hooks/useSelectedTaskId.test.tsx
@@ -1,8 +1,5 @@
 import type { ReactNode } from 'react';
 import React from 'react';
-import { useSelectedFormLayoutName } from './';
-import { layout1NameMock } from '../testing/layoutMock';
-import { layoutSet1NameMock } from '../testing/layoutSetsMock';
 import { renderHook } from '@testing-library/react';
 import type { ServicesContextProps } from 'app-shared/contexts/ServicesContext';
 import { ServicesContextProvider } from 'app-shared/contexts/ServicesContext';
@@ -11,10 +8,13 @@ import { MemoryRouter } from 'react-router-dom';
 import { createQueryClientMock } from 'app-shared/mocks/queryClientMock';
 import type { QueryClient } from '@tanstack/react-query';
 import { QueryKey } from 'app-shared/types/QueryKey';
+import { TASKID_FOR_STATELESS_APPS } from 'app-shared/constants';
 import { app, org } from '@studio/testing/testids';
+import { useSelectedTaskId } from './useSelectedTaskId';
 
 // Test data:
-const selectedLayoutSet = layoutSet1NameMock;
+export const layoutSet1NameMock = 'test-layout-set';
+export const layout1NameMock = 'Side1';
 
 const mockSetSearchParams = jest.fn();
 const mockSearchParams = { layout: layout1NameMock };
@@ -51,28 +51,33 @@ const wrapper = ({
   );
 };
 
-describe('useSelectedFormLayoutName', () => {
+describe('useSelectedTaskId', () => {
   afterEach(jest.clearAllMocks);
 
-  it('should return undefined when the select layout is invalid', async () => {
-    const { result } = renderHook(() => useSelectedFormLayoutName(selectedLayoutSet), { wrapper });
-    expect(result.current.selectedFormLayoutName).toEqual(undefined);
+  it('should return the default task id when it does not exist', async () => {
+    const { result } = renderHook(() => useSelectedTaskId(layoutSet1NameMock), { wrapper });
+    expect(result.current).toEqual(TASKID_FOR_STATELESS_APPS);
   });
 
-  it('should return selected layout when the selected layout is valid', async () => {
+  it('should return the selected task when it exists', async () => {
+    const taskTest = 'Task_Test';
+
     const client = createQueryClientMock();
-    client.setQueryData([QueryKey.FormLayoutSettings, org, app, selectedLayoutSet], {
-      pages: {
-        order: [layout1NameMock],
-      },
+    client.setQueryData([QueryKey.LayoutSets, org, app], {
+      sets: [
+        {
+          id: layoutSet1NameMock,
+          tasks: [taskTest],
+        },
+      ],
     });
 
-    const { result } = renderHook(() => useSelectedFormLayoutName(selectedLayoutSet), {
+    const { result } = renderHook(() => useSelectedTaskId(layoutSet1NameMock), {
       wrapper: ({ children }) => {
         return wrapper({ children, client });
       },
     });
 
-    expect(result.current.selectedFormLayoutName).toEqual(layout1NameMock);
+    expect(result.current).toEqual(taskTest);
   });
 });
